EntryItem 을 domain 패키지로 이동, move() 메서드로 내부 상태가 변경되기 때문에 ValueObject 로 볼 수 없다.
