+++ src/main/java/racing/domain/Entry.java
@@ -1,9 +1,7 @@
 package racing.domain;
 
-import racing.vo.EntryItem;
 import racing.vo.RacingRecord;
 
-import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
