+++ src/main/java/racing/domain/EntryItem.java
@@ -1,6 +1,4 @@
-package racing.vo;
-
-import racing.domain.RacingCar;
+package racing.domain;
 
 public class EntryItem implements Comparable<EntryItem>{
